docs(stories): keep old url version and enhance organization
