+++ src/stories/venmo/VenmoButton.stories.tsx
@@ -1,10 +1,10 @@
 import React, { FC } from "react";
 import type { PayPalScriptOptions } from "@paypal/paypal-js/types/script-options";
-import type { Story } from "@storybook/react";
 
-import { PayPalScriptProvider, FUNDING, PayPalButtons } from "../index";
-import { getOptionsFromQueryString, generateRandomString } from "./utils";
-import { InEligibleError, generateDocPageStructure } from "./commons";
+import { PayPalScriptProvider, FUNDING, PayPalButtons } from "../../index";
+import { getOptionsFromQueryString, generateRandomString } from "../utils";
+import { InEligibleError } from "../commons";
+import overrideStories from "./code";
 
 const scriptProviderOptions: PayPalScriptOptions = {
     "client-id":
@@ -16,6 +16,7 @@ const scriptProviderOptions: PayPalScriptOptions = {
 };
 
 export default {
+    id: "example/VenmoButton",
     title: "PayPal/VenmoButton",
     parameters: {
         controls: { expanded: true },
@@ -97,40 +98,4 @@ export const Default: FC<{
     );
 };
 
-const getDefaultCode = (): string =>
-    `import { PayPalScriptProvider, PayPalButtons } from "@paypal/react-paypal-js";
-
-// Custom error component
-const InEligibleError: FC<{ text?: string }> = ({ text }) => (
-	<h3 style={{ color: "#dc3545", textTransform: "capitalize" }}>
-		{text || "The component is ineligible to render"}
-	</h3>
-);
-
-export default function App() {
-	return (
-		<PayPalScriptProvider
-			options={{
-				"client-id":
-					"AdLzRW18VHoABXiBhpX2gf0qhXwiW4MmFVHL69V90vciCg_iBLGyJhlf7EuWtFcdNjGiDfrwe7rmhvMZ",
-				components: "buttons,funding-eligibility",
-				"enable-funding": "venmo",
-			}}
-		>
-			<PayPalButtons
-				fundingSource="venmo"
-				style={{
-                    "color": "blue"
-                }}
-			>
-				<InEligibleError text="You are not eligible to pay with Venmo." />
-			</PayPalButtons>
-		</PayPalScriptProvider>
-	);
-}`;
-
-(Default as Story).parameters = {
-    docs: {
-        page: () => generateDocPageStructure(getDefaultCode()),
-    },
-};
+overrideStories();
